Move install function into sprockets_amqp.web.
