+++ docs/api.rst
@@ -1,7 +1,5 @@
 API
 ===
-.. automodule:: sprockets.mixins.amqp
-   :members:
 
 AMQP Client
 -----------
